flatpickr
переместил константы с форматами, добавил настройки flatpickr по умолчанию, применил minDate
